% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_plot.R
\name{ag_gene_data}
\alias{ag_gene_data}
\title{Sub-setting Genome Annotation data to Focal Zoom Window}
\usage{
ag_gene_data(z_chr, z_start, z_end, n_tracks = 3, focal_genes = c())
}
\arguments{
\item{z_chr}{string, focal scaffold (eg. "mscaf_a1_01")}

\item{z_start}{integer, zoom starting position on scaffold (bp)}

\item{z_end}{integer, zoom end position on scaffold (bp)}

\item{n_tracks}{integer, number of gene tracks}

\item{focal_genes}{vector of strings, list of gene names to be tagges as "focal"}
}
\description{
Sub-setting Genome Annotation data to Focal Zoom Window
}
