% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_plot.R
\name{ag_plot_zoom}
\alias{ag_plot_zoom}
\title{Adds Genome Annotation Layers to a ggplot}
\usage{
ag_plot_zoom(
  z_chr,
  z_start,
  z_end,
  n_tracks = 3,
  focal_genes = c(),
  track_width = 0.5,
  label_offset = -0.4,
  labels = "all",
  style = "exons",
  arrow_size = ggplot2::unit(3, "pt")
)
}
\arguments{
\item{z_chr}{string, focal scaffold (eg. "mscaf_a1_01")}

\item{z_start}{integer, zoom starting position on scaffold (bp)}

\item{z_end}{integer, zoom end position on scaffold (bp)}

\item{n_tracks}{integer, number of gene tracks}

\item{focal_genes}{vector of strings, list of gene names to be tagged as "focal"}

\item{track_width}{double, width of exon boxes}

\item{label_offset}{double, horizontal offset of gene labels}

\item{labels}{string, toggles which genes are labelled. One of (\code{"all"}: all genes, \code{"focal"} focal genes only, \code{NA}: labels off)}

\item{style}{string, toggles plotting style. One of (\code{exons}: exons included, \code{simple}: gene extends and direction are plotted)}

\item{arrow_size}{double, size of gene arrow heads in \code{"simple"} style}
}
\description{
Adds Genome Annotation Layers to a ggplot
}
