% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arcgaz_exons_old_names}
\alias{arcgaz_exons_old_names}
\title{Antarctic Fur Seal Annotation Exon Coordinates (Old Names).}
\format{
A tibble with 166,497 rows and 12 variables:
\describe{
\item{gene_id}{string, gene ID, eg ("ANN06764")}
\item{gene_name}{string, gene name, eg ("RABL2A")}
\item{chr}{string, scaffold ID, eg ("mscaf_a1_01")}
\item{start}{integer, start position of the exon on the scaffold (bp)}
\item{end}{integer, end position of the exon on the scaffold (bp)}
\item{width}{integer, width of the exon (bp)}
\item{strand}{string, +/- indicator if the gene annotation is with respect to the forward(+)/reverse(-) strand. }
\item{exon_id}{double, exon ID, eg ("5892")}
\item{exon_name}{string, exon name, eg ("ANN06764-RA:exon:240")}
\item{gstart}{integer, genomic start position of the exon (bp)}
\item{gend}{integer, genomic end position of the exon (bp)}
\item{gmid}{integer, central genomic position of the exon, useful for labels (bp)}
}
}
\usage{
arcgaz_exons_old_names
}
\description{
A data set containing the coordinates for all exons within the annotation for the A. gazella reference genome.
Version with the initial names for the unplaced scaffolds (eg. \verb{ScDpJTZ_94;HRSCAF=120} instead of \code{ScDpJTZ_94_HRSCAF_120}).
}
\keyword{datasets}
