% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arcgaz_genes}
\alias{arcgaz_genes}
\title{Antarctic Fur Seal Annotation Gene Coordinates.}
\format{
A tibble with 23,408 rows and 9 variables:
\describe{
\item{gene_id}{string, gene ID, eg ("ANN06764")}
\item{gene_name}{string, gene name, eg ("RABL2A")}
\item{chr}{string, scaffold ID, eg ("mscaf_a1_01")}
\item{start}{integer, start position of the gene on the scaffold (bp)}
\item{end}{integer, end position of the gene on the scaffold (bp)}
\item{strand}{string, +/- indicator if the gene annotation is with respect to the forward(+)/reverse(-) strand. }
\item{gstart}{integer, genomic start position of the gene (bp)}
\item{gend}{integer, genomic end position of the gene (bp)}
\item{gmid}{integer, central genomic position of the gene, useful for labels (bp)}
}
}
\usage{
arcgaz_genes
}
\description{
A data set containing the coordinates for all genes within the annotation for the A. gazella reference genome.
}
\keyword{datasets}
