% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arcgaz_genome}
\alias{arcgaz_genome}
\title{Antarctic Fur Seal Genome Scaffold Overview.}
\format{
A tibble with 18 rows and 6 variables:
\describe{
\item{chr}{string, scaffold ID, eg ("mscaf_a1_01")}
\item{length}{integer, length of the scaffold (bp)}
\item{start}{integer, genomic start position of the scaffold (bp)}
\item{end}{integer, genomic end position of the scaffold (bp)}
\item{mid}{integer, central genomic position of the scaffold, useful for labels (bp)}
\item{eo}{integer, 0/1 indicator of even(1)/odd(0) scaffold type. }
}
}
\usage{
arcgaz_genome
}
\description{
A data set containing the coordinates for the main 18 scaffolds in the A. gazella reference genome.
}
\keyword{datasets}
