% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arcgaz_genome_start_old_names}
\alias{arcgaz_genome_start_old_names}
\title{Antarctic Fur Seal Genome Scaffold Start Positions (Old Names).}
\format{
A tibble with 18 rows and 2 variables:
\describe{
\item{chr}{string, scaffold ID, eg ("mscaf_a1_01")}
\item{start}{integer, genomic start positions of the scaffold (bp)}
}
}
\usage{
arcgaz_genome_start_old_names
}
\description{
A data set containing the genomic start positions of the scaffolds of the A. gazella reference genome.
Version with the initial names for the unplaced scaffolds (eg. \verb{ScDpJTZ_94;HRSCAF=120} instead of \code{ScDpJTZ_94_HRSCAF_120}).
}
\keyword{datasets}
