% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argcaz_busco}
\alias{argcaz_busco}
\title{Antarctic Fur Seal BUSCO Coordinates.}
\format{
A tibble with 15,663 rows and 17 variables:
\describe{
\item{busco_id}{string, BUSCO ID, eg ("3at33554")}
\item{status}{string, BUSCO status (one of "Complete", "Missing", "Fragmented", "Duplicated")}
\item{chr}{string, scaffold ID, eg ("mscaf_a1_01")}
\item{start}{integer, start position of the BUSCO on the scaffold (bp)}
\item{end}{integer, end position of the BUSCO on the scaffold (bp)}
\item{dstart}{integer, strand aware start position of the BUSCO on the scaffold (bp)}
\item{dend}{integer, strand aware end position of the BUSCO on the scaffold (bp)}
\item{gstart}{integer, genomic start position of the BUSCO (bp)}
\item{gend}{integer, genomic end position of the BUSCO (bp)}
\item{gmid}{integer, central genomic position of the BUSCO, useful for labels (bp)}
\item{gdstart}{integer, strand aware genomic start position of the BUSCO (bp)}
\item{gdend}{integer, strand aware genomic end position of the BUSCO (bp)}
\item{strand}{string, +/- indicator if the BUSCO direction is forward(+) or reverse(-). }
\item{score}{BUSCO score. }
\item{length}{double, BUSCO length (bp)}
\item{orthodb_url}{string, link to the ortho db entry for the BUSCO}
\item{description}{string, verbal BUSCO description}
}
}
\usage{
argcaz_busco
}
\description{
A data set containing the coordinates and scores of all BUSCOS found for the A. gazella reference genome.
The is based on the \code{full_table.tsv} summary of the BUSCO report based on the carnivora reference set.
}
\keyword{datasets}
