% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argcaz_busco_go}
\alias{argcaz_busco_go}
\title{Antarctic Fur Seal BUSCO to GO-term translation.}
\format{
A tibble with 9,471 rows and 2 variables:
\describe{
\item{busco_id}{string, BUSCO ID, eg ("3at33554")}
\item{go_terms}{string, comma separated list of GO terms)}
}
}
\usage{
argcaz_busco_go
}
\description{
A data set containing the GO terms ("biological process") associated with each BUSCO.
Only BUSCOS with the status \code{Complete} are listed.
Go terms were harvested from "https://www.ebi.ac.uk/QuickGO/" on the 2023-11-22s.
}
\keyword{datasets}
