% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{directed_coords}
\alias{directed_coords}
\title{Helper to order strand-specific coordinates of genomic features}
\usage{
directed_coords(dat)
}
\arguments{
\item{dat}{a \code{data.frame} containing the columns \code{start}, \code{end} and \code{strand}. Strand needs to be in ("+"/"-" format)}
}
\description{
This function flips the \code{start} and \code{end} coordinate of a genomic feature depending on the feature's strand.
}
