% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_bg.R
\name{geom_ag_genome}
\alias{geom_ag_genome}
\title{ggplot geom for genome background layer}
\usage{
geom_ag_genome(fills = c("transparent", rgb(0, 0, 0, 0.1)))
}
\arguments{
\item{fills}{a color vector with the length of two (default \code{c("transparent", rgb(0,0,0,0.1))}) for the alternating scaffold shading. First value is for even, second for odd scaffolds.}
}
\description{
ggplot geom for genome background layer
}
\examples{
ggplot2::ggplot() +
 geom_ag_genome() +
 scale_x_ag_genome() +
 ggplot2::coord_cartesian(xlim = c(0, max(arcgazgen::arcgaz_genome$end)),
                          ylim = 0:1,
                          expand = 0) +
 theme_ag_genome()
}
