% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_bg.R
\name{scale_x_ag_genome}
\alias{scale_x_ag_genome}
\title{ggplot x scale for genome plots}
\usage{
scale_x_ag_genome(name = "Genomic Position (Gb)", name_top = "Scaffold ID")
}
\arguments{
\item{name}{Title for the primary x axis}

\item{name_top}{Title for the secondary x axis}
}
\description{
ggplot x scale for genome plots
}
