% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_bg.R
\name{theme_ag_genome}
\alias{theme_ag_genome}
\title{ggplot theme for genome plots}
\usage{
theme_ag_genome(..., lwd = 0.3)
}
\arguments{
\item{...}{parameters passed on to \code{theme_minimal()}}

\item{lwd}{the \code{linewidth} for axis lines and ticks (default: \code{0.3})}
}
\description{
ggplot theme for genome plots
}
